# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /racks:
    get:
      operationId: listRacks
      tags: [racks]
      summary: List racks
      responses:
        "200":
          description: List of racks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rack"
    post:
      operationId: createRack
      tags: [racks]
      summary: Create a rack
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RackInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rack"
  /racks/{id}:
    patch:
      operationId: updateRack
      tags: [racks]
      summary: Update a rack
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RackInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rack"
    delete:
      operationId: deleteRack
      tags: [racks]
      summary: Delete a rack
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /members:
    get:
      operationId: listMembers
      tags: [members]
      summary: List members
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: className
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of members
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberListResponse"
    post:
      operationId: createMember
      tags: [members]
      summary: Create a member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
  /members/{id}:
    patch:
      operationId: updateMember
      tags: [members]
      summary: Update a member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
    delete:
      operationId: deleteMember
      tags: [members]
      summary: Delete a member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /favorites:
    get:
      operationId: getFavorites
      tags: [favorites]
      summary: Get user favorites
      responses:
        "200":
          description: Favorite books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Book"
    post:
      operationId: addFavorite
      tags: [favorites]
      summary: Add a book to favorites
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FavoriteInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /favorites/{bookId}:
    delete:
      operationId: removeFavorite
      tags: [favorites]
      summary: Remove from favorites
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /staff:
    get:
      operationId: listStaff
      tags: [staff]
      summary: List staff
      responses:
        "200":
          description: List of staff
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffMember"
    post:
      operationId: createStaff
      tags: [staff]
      summary: Create staff
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
  /staff/{id}:
    patch:
      operationId: updateStaff
      tags: [staff]
      summary: Update staff
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
    delete:
      operationId: deleteStaff
      tags: [staff]
      summary: Delete staff
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    Book:
      type: object
      required: [id, isbn, title, author, stock, status]
      properties:
        id:
          type: integer
        isbn:
          type: string
        title:
          type: string
        author:
          type: string
        publisher:
          type: ["string", "null"]
        year:
          type: ["integer", "null"]
        stock:
          type: integer
        available:
          type: integer
        description:
          type: ["string", "null"]
        pages:
          type: ["integer", "null"]
        categoryId:
          type: ["integer", "null"]
        rackId:
          type: ["integer", "null"]
        coverUrl:
          type: ["string", "null"]
        categoryName:
          type: ["string", "null"]
        rackName:
          type: ["string", "null"]
        borrowCount:
          type: integer
        status:
          type: string
          enum: [available, borrowed]
        createdAt:
          type: ["string", "null"]
...
    Rack:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        location:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        bookCount:
          type: integer
        createdAt:
          type: ["string", "null"]
    RackInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        location:
          type: string
        description:
          type: string
    Member:
      type: object
      required: [id, memberNumber, name, status]
      properties:
        id:
          type: integer
        memberNumber:
          type: string
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        className:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive]
        fine:
          type: number
        borrowCount:
          type: integer
        createdAt:
          type: ["string", "null"]
    MemberListResponse:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Member"
        total:
          type: integer
    MemberInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        className:
          type: string
    MemberUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        className:
          type: string
        status:
          type: string
          enum: [active, inactive]
...
    FavoriteInput:
      type: object
      required: [bookId]
      properties:
        bookId:
          type: integer
    StaffMember:
      type: object
      required: [id, username, name, role]
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
        createdAt:
          type: ["string", "null"]
    StaffInput:
      type: object
      required: [username, password, name, role]
      properties:
        username:
          type: string
        password:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
    StaffUpdate:
      type: object
      properties:
        password:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
...